requirements.txt
streamlit
pandas
numpy 
plotly
biopython 
py3Dmol
matplotlib 
stmol
ipython_genutils